Add files via upload
This project is motivated by the increase in popularity of Sudoku and similar logic
puzzles over the past decade . Numerous serial Sudoku solver implementations
exist and are readily available. However, due to its recency, there is relatively little
work on parallel implementations. Parallelizing a serial Sudoku solver can improve
its speed and increase the viability of solving larger size Sudokus. The Sudoku
solver algorithm also conveys some of the basic tradeoffs of parallel software
development such as dependencies and work-sharing which are interesting to
inspect and study. We propose that a method for solving the puzzle quickly will be derived that takes
advantage of many core architecture on modern and future computer systems
which will bring down the time cost of solving an NxN Sudoku to a reasonable
amount.
Our goal is to allow for the algorithm to take advantage of the many-core
architecture to further reduce its running time. This is very important and is at the
h